Fixed script exit status when uncrustify fails.
cleanup() was called correctly, but it cannot terminate the whole script
because it is called from a subshell. It can only terminate its own
subshell. The fix works because there is only one subshell to take care
of, but the whole approach is fragile. There ought to be a better way!